Agents/api_requirements.txt
# Synapse Protocol API Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0
python-multipart==0.0.6
cors==1.0.1
python-dotenv==1.0.0
asyncio-mqtt==0.16.1

# Existing requirements
uagents==0.12.0
yfinance==0.2.18
pandas==2.0.3
numpy==1.24.3
scipy==1.11.1
statsmodels==0.14.0
requests==2.31.0